requirements.txt
nltk
pandas
scikit-learn
spacy
google-genai
python-dotenv
pydantic
joblib
scipy
sentence-transformers
